+++ app/src/main/java/io/github/yawnoc/strokeinput/StrokeSequenceBar.java
@@ -1,5 +1,5 @@
 /*
-  Copyright 2021, 2023 Conway
+  Copyright 2021, 2023, 2025 Conway
   Licensed under the GNU General Public License v3.0 (GPL-3.0-only).
   This is free software with NO WARRANTY etc. etc.,
   see LICENSE or <https://www.gnu.org/licenses/>.
@@ -25,7 +25,7 @@ public class StrokeSequenceBar
 
   public void setStrokeDigitSequence(final String strokeDigitSequence)
   {
-    if (strokeDigitSequence.length() > 0)
+    if (!strokeDigitSequence.isEmpty())
     {
       final String strokeSequence =
               strokeDigitSequence